pkg: move version package into pkg
The code is used by the operator, but other than that has nothing to
do with it. Therefore "pkg" is a better place.
