+++ pkg/k8sutil/client.go
@@ -11,7 +11,7 @@ import (
 	"regexp"
 	"strconv"
 
-	"github.com/intel/pmem-csi/pkg/pmem-csi-operator/version"
+	"github.com/intel/pmem-csi/pkg/version"
 	"k8s.io/client-go/discovery"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
